Move numberFromEnv to a playwright-free module so the test runs in CI
The test imported `numberFromEnv` from `profiles.ts`, which hard-imports `playwright` at module
load. CI installs only the root workspace, not `agent-computer`, so that import throws `Cannot find
package 'playwright'`, drops the file's tests, and fails the `tests` and `verify` jobs.

`numberFromEnv` moves to its own `agent-computer/src/env.ts`, which imports nothing; `profiles.ts`
re-exports it so `index.ts` and the browser-limit constants are untouched, and the test imports it
from `./env`. Verified with `agent-computer/node_modules` moved aside — a runner without playwright —
where `bun test` now passes 5/5 instead of failing to import.
